requirements.txt
certifi==2019.3.9
chardet==3.0.4
colorama==0.4.1
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.1
django-extra-fields==1.2.3
django-taggit==1.1.0
django-taggit-autocomplete-modified==0.1.0b6
djangorestframework==3.9.4
httpie==1.0.2
idna==2.8
itypes==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
Pillow==6.0.0
Pygments==2.4.0
pytz==2019.1
requests==2.22.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.2
virtualenv==16.6.0
